requirements.txt
fastapi~=0.121.2
uvicorn[standard]~=0.38.0
openai~=2.8.1
faiss-cpu~=1.13.0
numpy~=2.3.5
python-dotenv~=1.2.1
pydantic~=2.12.4
sentence-transformers~=5.1.2
torch~=2.9.1
scikit-learn~=1.7.2
requests~=2.32.5

pillow~=12.0.0
h11~=0.16.0
pip~=25.1.1
distro~=1.9.0
filelock~=3.20.0
PyYAML~=6.0.3
sniffio~=1.3.1
typing_extensions~=4.15.0
scipy~=1.16.3
packaging~=25.0
click~=8.3.1
httpcore~=1.0.9
idna~=3.11
httpx~=0.28.1
jiter~=0.12.0
charset-normalizer~=3.4.4
setuptools~=80.9.0
regex~=2025.11.3
threadpoolctl~=3.6.0
mpmath~=1.3.0
sympy~=1.14.0
tqdm~=4.67.1
triton~=3.5.1
networkx~=3.5
urllib3~=2.5.0
fsspec~=2025.10.0
ujson~=5.11.0
MarkupSafe~=3.0.3
joblib~=1.5.2
anyio~=4.11.0
pydantic_core~=2.41.5
certifi~=2025.11.12
starlette~=0.49.3
orjson~=3.11.4
annotated-types~=0.7.0
Jinja2~=3.1.6
tokenizers~=0.22.1
safetensors~=0.6.2
transformers~=4.57.1
huggingface-hub~=0.36.0